Rename and move arund multiple widgets
