+++ app/src/components/feature/widgets/Widget.component.jsx
