+++ app/src/components/feature/widgets/graph/GraphStats.component.jsx
@@ -20,7 +20,7 @@ import 'overlayscrollbars/styles/overlayscrollbars.css';
 import PropTypes from 'prop-types';
 import { useContext, useEffect, useState } from 'react';
 import { RootStoreContext } from 'stores/RootStore';
-import ChartAlertComponent from '../chart/ChartAlert.component';
+import ChartAlertComponent from '../WidgetAlert.component';
 
 function GraphStats(props) {
     const store = useContext(RootStoreContext);
